+++ common/src/main/kotlin/me/mochibit/createharmonics/audio/instance/SimpleTickableSoundInstance.kt
@@ -19,7 +19,7 @@ class SimpleTickableSoundInstance(
     volumeSupplier: () -> Float = { 1.0f },
     pitchSupplier: () -> Float = { 1.0f },
     posSupplier: () -> BlockPos = { BlockPos.ZERO },
-    radiusSupplier: () -> Int = { 64 },
+    radiusSupplier: () -> Float = { 64f },
 ) : SuppliedSoundInstance(
         soundEvent,
         soundSource,
