+++ forge/src/main/kotlin/me/mochibit/createharmonics/audio/instance/SimpleShipStreamSoundInstance.kt
@@ -20,7 +20,7 @@ class SimpleShipStreamSoundInstance(
     private val ship: Ship,
     volumeSupplier: () -> Float = { 1.0f },
     pitchSupplier: () -> Float = { 1.0f },
-    radiusSupplier: () -> Int = { 64 },
+    radiusSupplier: () -> Float = { 64f },
     randomSource: RandomSource = RandomSource.create(),
     soundSource: SoundSource = SoundSource.RECORDS,
     looping: Boolean = false,
